Create initial data model for TidePrediction
This change introduces the `model` module, and creates a
`TidePrediction` struct in it. The `TidePrediction` currently has two
fields: a `f32` for the tide reading, and a
`chrono::DateTime<FixedOffset>` for the date. This requires us to take a
dependency on `chrono`, which seems like "the good DateTime library for
rust".

In the `home_page` function, we now create a dummy `TidePrediction`, and
print its `Debug` implementation.