Implement TidePredictionPair finding algorithm and Display
This commit introduces major new functionality for WTIIRN: the ability
to know WHICH WAY THE TIDE IS GOING!

This major feature improvement is implemented by introducing a new
struct, the `TidePredictionPair`, along with the algorithm to generate
it from a list of tides and a time.

By determining the most recent and next tide predictions, we can
determine the directionality of the tide, and display it to the user. By
implementing the `fmt::Display` trait for both `TidePrediction`s and
`TidePredictionPair`s, we can now render these stucts much more nicely.

Along the way, I also decided to seperate out the algorithms into their
own module, `compute`.