Add great circle distance calculation
This commit adds a model for representing points on the surface of the
Earth using a simplified spherical model (not the more accurate but
trickier oblate spheroid).

Since we're on a modern 64 bit system, we can use the simpler formulas
offered here:

https://en.wikipedia.org/wiki/Great-circle_distance

Additionally, the model for representing points can transform its
lattitude and longitude data from degrees to radians (which is
necessary because Rust's stdlib trigonometric operations operate on
radians, but latitude and longitude are most commonly expressed in
degrees).