Adopt uom crate for Measurements
This commit adds the uom crate, and uses it to replace our custom type
alias for `Meters`. This allows us to easily display the distance from
point atkinson in Kilometers rather than meters, and in the future will
allow us to work with NOAA data (in Feet) alongside our canadian data in
Meters.