Initial implementation of predictions_for_station
This API feels a little weird right now, but we think it might feel more
natural once we're actually using it. Or maybe it won't! In that case,
we'll change it once we know what we're actually doing.

We're storing predictions in a `HashMap`, keyed to station ids (`u64`)
which seems fine.

We also expose a method to find the predictions associated with the given station,
and return them.