Fix the build
Unfortunately, I was moving a little too quick, and broke the master
build. It is important to `use` the types you want to use in your
structs! Also: probably a good idea not to use numerous references in
struct fields, unless you want to write lifetime parameters all over the
place.

On the plus side, I've now made it so you can't merge to master without
CI passing, which will prevent this kind of problem in the future!