Add explanation of Chart Datum to prediction
This commit adds an explanation of the jargon term 'Chart Datum' to
the implementation of `fmt::Display` for `TidePrediction`. which is
eventually rendered on the home page. In the future, it may make sense
to perform this task with a templating system instead, and have the
linked explanation in the homepage template rather than associating it
with the `TidePrediction`.