First try at http -> https redirection
This change uses the `x-forwarded-proto` header to check if heroku has
transmited the request from http, and returns a 301 redirect to the
https url in that case.

In order to test this, I extracted the route handler to a function,
which means we can now unit test our routes very easily.

I've been able to manually test that the 301 is returned locally, but
I'm still not 100% confident this will work how we expect. Fingers
crossed!