Interpolation between tide predictions
This commit implements interpolation between tide predictions.
Given a pair of adjacent tide predictions, the function in this
commit calculates the approximate tide level at a the given time
between them.

The function assumes that between two predictions, the tide level will
follow a sinusoidal curve with critical points at each predictions (i.e.
the tide's rate of change goes to zero at low or high tide). This
assumption may not be very good for certain stations, but it's a
defensible approximation.

The function is also susceptible to misuse. It doesn't, for example,
check that the given tide predictions are within some interval of each
other (24 hours would be a sensible upper bound) or that the given
time is actually between the given tide preidctions. I'm not sure if
those constraints could actually be encoded in the type system, and they
would certainly complicate the function's signature. As such, the checks
would have to be performed at run-time, and the function would return a
result. An alternative might be to take a `TidePair` rather than a pair
of tide predictions.

Rather than do that complicated work, I elected to push this version,
which I will try to integrate into the system elsewhere and see what API
changes (if any) suggest themselves to make it more difficult to misuse.