Display current tide status
This change updates the homepage to display what I'm calling the
"current status" of the tide. This is a natural language representation
of what we expect users actually want to know:

If the tide is going out, then how much more will it go down, and for
how long.

This change also updates the previous "detail" display to instead show a
table, which we believe is a more useful form for the information it
contains.